+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.8.0] - 2020-12-14
+
+### Added
+
+- exports for `CustomEntityDetectionModel`
+- support for `_$text` and `_$entity` at quick replies
+
+### Fixed
+
+- examples in docs
 ## [3.7.5] - 2020-12-01
 
 ### Fixed